Extract meta
